framework:add safer method to increment m_buff_ptr
Currently, a known bug causes the radio agent threads to segfault and
die as a result of receiving an M1 message when expecting an M2 message,
the agents attempt to parse an M2 TLV but if they end up receiving an M1
message.
Upon inspection, this is caused when `m_buff_ptr` is incremented by
incorrect amounts as a result of the mismatch, causing a segmentation
fault.
This added method verifies that the incrementation does not exceed the
allocated buffer length before anything is done, and returns false on
failure, causing the process to halt before segfaulting.

Signed-off-by: Gal Wiener <gal.wiener@intel.com>